Supports ephemeralWebBrowserSession on iOS 13
Hello, 

Actually, on iOS 12+, when you login, the cookie is kept until the app is killed. 

Since iOS 13 it's possible to make session ephemeral. 
With this option prefersEphemeralWebBrowserSession, the cookie is killed when tab browser is closed. 

Pros : 
- no need to logout manualy users (for account switch for example) 
Cons : 
- user should reconnect each time the tab is presented.

It fix https://github.com/proyecto26/react-native-inappbrowser/issues/76 , on iOS 13.

TODO : Improvement : make this parameter as an option 

Best Regards